Powerups

+++ Eclipse/Jumpy/src/com/example/jumpy/Powerup.java
@@ -3,9 +3,9 @@ package com.example.jumpy;
 public class Powerup extends Item
 {
 	private int type;
-	private int value;
+	private double value;
 	
-	public Powerup(int id, String name, String description, boolean multiple, int price, int image, int quantity, int type, int value)
+	public Powerup(int id, String name, String description, boolean multiple, int price, int image, int quantity, int type, double value)
 	{
 		super(id, name, description, multiple, price, image, quantity);
 		this.value = value;
@@ -17,7 +17,7 @@ public class Powerup extends Item
 		return type;
 	}
 	
-	public int getValue()
+	public double getValue()
 	{
 		return value;
 	}
